rename sb InstanceDTO file to match class name
